+++ bauh/gems/appimage/__init__.py
@@ -9,7 +9,6 @@ from bauh.commons import resource
 ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
 APPIMAGE_SHARED_DIR = f'{SHARED_FILES_DIR}/appimage'
 INSTALLATION_DIR = f'{APPIMAGE_SHARED_DIR}/installed'
-SUGGESTIONS_FILE = f'https://raw.githubusercontent.com/vinifmor/{__app_name__}-files/master/appimage/suggestions.txt'
 CONFIG_FILE = f'{CONFIG_DIR}/appimage.yml'
 APPIMAGE_CONFIG_DIR = f'{CONFIG_DIR}/appimage'
 UPDATES_IGNORED_FILE = f'{APPIMAGE_CONFIG_DIR}/updates_ignored.txt'
@@ -19,8 +18,6 @@ APPIMAGE_CACHE_DIR = f'{CACHE_DIR}/appimage'
 DATABASE_APPS_FILE = f'{APPIMAGE_CACHE_DIR}/apps.db'
 DATABASE_RELEASES_FILE = f'{APPIMAGE_CACHE_DIR}/releases.db'
 DATABASES_TS_FILE = f'{APPIMAGE_CACHE_DIR}/dbs.ts'
-SUGGESTIONS_CACHED_FILE = f'{APPIMAGE_CACHE_DIR}/suggestions.txt'
-SUGGESTIONS_CACHED_TS_FILE = f'{APPIMAGE_CACHE_DIR}/suggestions.ts'
 DOWNLOAD_DIR = f'{TEMP_DIR}/appimage/download'