+++ bauh/gems/flatpak/__init__.py
@@ -9,7 +9,6 @@ from bauh.api.paths import CONFIG_DIR
 from bauh.commons import resource
 
 ROOT_DIR = os.path.dirname(os.path.abspath(__file__))
-SUGGESTIONS_FILE = f'https://raw.githubusercontent.com/vinifmor/{__app_name__}-files/master/flatpak/suggestions.txt'
 CONFIG_FILE = f'{CONFIG_DIR}/flatpak.yml'
 FLATPAK_CONFIG_DIR = f'{CONFIG_DIR}/flatpak'
 UPDATES_IGNORED_FILE = f'{FLATPAK_CONFIG_DIR}/updates_ignored.txt'